+++ .storybook/preview.jsx
@@ -2,18 +2,13 @@ import React from 'react';
 import {addons} from 'storybook/preview-api';
 import {UPDATE_GLOBALS, STORY_ARGS_UPDATED} from "storybook/internal/core-events";
 
-import './preview.scss';
-
 // Copy imports from src/index.ts, in the same order
-import '../src/globals/reset.module.scss';
+import '../src/globals/reset.scss';
 import '../src/globals/_variables.scss';
 import '../src/tokens/spacings/spacings.scss';
 import '../src/tokens/colors/colors.scss';
 import '../src/tokens/borders/borders.scss';
 import '@fontsource-variable/nunito-sans';
-
-// Uncomment to use legacy css in storybook
-// import '../dist/legacy-global-bundle.css';
 
 // Uncomment to use legacy css in storybook
 // import '../dist/legacy-global-bundle.css';