+++ src/components/Collapsible/ControlledCollapsible.tsx
@@ -3,7 +3,7 @@ import clsx from 'clsx';
 import type {ControlledCollapsibleProps} from './Collapsible.types';
 import {Typography} from '~/components';
 import {ChevronRight} from '~/icons';
-import {layout, reset} from '~/globals/css-utils.js';
+import {layout} from '~/globals/css-utils.js';
 import styles from './Collapsible.module.scss';
 
 const ControlledCollapsibleForwardRef: React.ForwardRefRenderFunction<HTMLDivElement, ControlledCollapsibleProps> = ({
@@ -16,7 +16,6 @@ const ControlledCollapsibleForwardRef: React.ForwardRefRenderFunction<HTMLDivEle
     ...other
 }, ref) => {
     const classNameProps = clsx(
-        reset,
         ['moonstone-collapsible', styles['moonstone-collapsible']],
         className
     );