+++ src/components/Switch/ControlledSwitch.tsx
@@ -2,7 +2,6 @@ import React, {MutableRefObject, useRef} from 'react';
 import clsx from 'clsx';
 import type {ControlledSwitchProps} from './Switch.types';
 import {onAccessibleClick} from '~/hooks';
-import {reset} from '~/globals/css-utils.js';
 import styles from './Switch.module.scss';
 
 const ControlledSwitchForwardRef: React.ForwardRefRenderFunction<HTMLDivElement, ControlledSwitchProps> = ({
@@ -19,7 +18,6 @@ const ControlledSwitchForwardRef: React.ForwardRefRenderFunction<HTMLDivElement,
         <div
             ref={ref}
             className={clsx(
-                reset,
                 ['moonstone-switch', styles['moonstone-switch']],
                 checked && ['moonstone-switch_checked', styles['moonstone-switch_checked']],
                 isDisabled && ['moonstone-switch_disabled', styles['moonstone-switch_disabled']],