+++ src/components/Textarea/ControlledTextarea.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import clsx from 'clsx';
 import type {ControlledTextareaProps} from './Textarea.types';
-import {reset} from '~/globals/css-utils.js';
 import styles from './Textarea.module.scss';
 
 export const ControlledTextarea = React.forwardRef<HTMLTextAreaElement, ControlledTextareaProps>(({
@@ -21,7 +20,6 @@ export const ControlledTextarea = React.forwardRef<HTMLTextAreaElement, Controll
         <textarea
             ref={ref}
             className={clsx(
-                reset,
                 ['moonstone-textarea', styles['moonstone-textarea']],
                 isResizable && ['moonstone-textarea_resizable', styles['moonstone-textarea_resizable']],
                 className